Move each APA to its own file
