+++ array_processing/algorithms/__init__.py
@@ -1,2 +1,3 @@
-from .array_algorithms import wlsqva, fk_freq
+from .wlsqva import wlsqva
+from .fk_freq import fk_freq
 from .helpers import wlsqva_proc, getrij
